+++ src/Brickables/OneTextColumn.php
@@ -6,13 +6,13 @@ use Okipa\LaravelBrickables\Abstracts\Brickable;
 
 class OneTextColumn extends Brickable
 {
-    protected function setStoreValidationRules(): array
+    public function validateStoreInputs(): array
     {
-        return ['text' => ['required', 'string']];
+        return request()->validate(['text' => ['required', 'string']]);
     }
 
-    protected function setUpdateValidationRules(): array
+    public function validateUpdateInputs(): array
     {
-        return ['text' => ['required', 'string']];
+        return request()->validate(['text' => ['required', 'string']]);
     }
 }